refactor(AuthenticatedLayout): move component to its own directory
