+++ src/components/router/Router.tsx
@@ -4,7 +4,7 @@ import UnauthenticatedLayout from '~components/layouts/UnauthenticatedLayout'
 import Login from '~/pages/login/Login'
 import ValidateAuthentication from '~components/auth/Authentication'
 import Home from '~/pages/home/Home'
-import AuthenticatedLayout from '~components/layouts/AuthenticatedLayout'
+import AuthenticatedLayout from '~/components/layouts/AuthenticatedLayout/AuthenticatedLayout'
 import { analysisRoutes } from './analysisRoutes'
 import { parametersRoutes } from './parametersRoutes'
 import { reportsRoutes } from './reportsRoutes'
